pyproject.toml
[tool.poetry]
name = "quotes api"
version = "0.1.0"
description = ""
authors = ["karelumair <umair.karel@sarvaha.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
python-dotenv = "^0.21.0"
flask-mongoengine = "^1.0.0"
flask-restful = "^0.3.9"
pre-commit = "^2.21.0"
pylint = "^2.15.10"
selenium = "^4.8.0"
beautifulsoup4 = "^4.11.2"
requests = "^2.28.2"
html5lib = "^1.1"
celery = "^5.2.7"
redis = "^4.5.4"
pydantic = "^1.10.4"
pytest = "^7.2.1"
pytest-celery = "^0.0.0"
flask-apscheduler = "^1.12.4"
flask-pydantic = "^0.11.0"
flask-limiter = "^3.3.0"
gunicorn = "^20.1.0"
flask-cors = "^3.0.10"
flask-jwt-extended = "^4.4.4"
black = "^23.3.0"

[tool.pylint.messages_control]
disable = "E1101, W0718"
extension-pkg-whitelist = "pydantic"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
